requirements.txt
aiogram
google-generativeai
flask